+++ cli/cmd/neverlauncher/main.go
@@ -11,7 +11,7 @@ import (
 	"time"
 )
 
-var version = "0.10.7"
+var version = "0.11.0"
 
 const cliSchemaVersion = "1.0"