+++ plugins/bridge-common/src/main/java/ru/neverlauncher/bridge/common/BridgeDefaults.java
@@ -1,7 +1,7 @@
 package ru.neverlauncher.bridge.common;
 
 public final class BridgeDefaults {
-    public static final String VERSION = "0.10.7";
+    public static final String VERSION = "0.11.0";
     public static final String VELOCITY_ID = "neverlauncher-velocity-bridge";
     public static final String PAPER_ID = "neverlauncher-paper-bridge";
     public static final String PURPUR_ID = "neverlauncher-purpur-bridge";